backend/requirements.txt
flask
flask-cors
pandas
openpyxl
python-docx
PyPDF2
psycopg2-binary
SQLAlchemy
Flask-SQLAlchemy
python-dotenv
